+++ CHANGELOG.md
@@ -9,6 +9,7 @@
   - Added support to Logical enclosure `:reconfigure` and `:set_script`
   - Added support to Network set actions
   - Added support to Datacenter actions
+  - Added support to Interconnect actions
 
 ### 0.1.1
   - Fixed Ruby SDK version to 1.0.0